refactor: create or replace statutory declaration v2 service
